+++ web/modules/custom/hoeringsportal_project/modules/hoeringsportal_project_fixtures/src/Fixture/ProjectCategoryFixture.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Drupal\hoeringsportal_project_fixtures\Fixture;
+
+use Drupal\hoeringsportal_base_fixtures\Fixture\TaxonomyTermFixture;
+
+/**
+ * Project category fixture.
+ *
+ * @package Drupal\hoeringsportal_project_fixtures\Fixture
+ */
+class ProjectCategoryFixture extends TaxonomyTermFixture {
+
+  /**
+   * {@inheritdoc}
+   */
+  protected static $vocabularyId = 'project_categories';
+
+  /**
+   * {@inheritdoc}
+   */
+  protected static $terms = [
+    'Byudvikling',
+    'Offentlig transport',
+  ];
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getGroups() {
+    return array_merge(parent::getGroups(), ['project']);
+  }
+
+}