+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 * [PR-398](https://github.com/itk-dev/hoeringsportal/pull/398)
   Added Deskpro test data
+* [pr-399](https://github.com/itk-dev/hoeringsportal/pull/399)
+  * Add content type Project subpage
+  * Update paragraphs for projects
+  * Add list view for project
 * [pr-394](https://github.com/itk-dev/hoeringsportal/pull/394)
   Add content type Project page